Created python function to erase list of big companies from a 2d list of all businesses in the list.
Assuming the input is a 2D list containing the name of the businesses and the URL linking to a google map of that business location, this function will return a list of similar format, but without any of the big businesses. Future modifications will include inputting the small businesses into the small business database.